+++ global/types.h
@@ -115,9 +115,6 @@ typedef struct {
 #define MAX_SHADE			(0x300)
 #define MAX_ROOMLIGHT_UNIT	(0x2000 / (WIBBLE_SIZE/2))
 
-// Maximum surface dimensions for DX5
-#define MAX_SURFACE_SIZE	(2048)
-
 // SW Renderer Detail Settings
 #define SW_DETAIL_LOW		(0 * 0x400 * W2V_SCALE)
 #define SW_DETAIL_MEDIUM	(3 * 0x400 * W2V_SCALE)
@@ -227,6 +224,12 @@ typedef struct {
 #define IFL_CODEBITS	(0x3E00)
 #define IFL_REVERSE		(0x4000)
 #define IFL_CLEARBODY	(0x8000)
+
+// Glow tint colors
+#define GLOW_FLARE_COLOR	(0xFF8080) // Flare
+#define GLOW_PISTOL_COLOR	(0x3F380F) // Pistol/Magnums/Uzi gunfire
+#define GLOW_M16_COLOR		(0x7F701F) // M16 gunfire
+#define GLOW_GUNSHOT_COLOR	(0x7F701F) // Skidoo/Enemy gunfire
 
 /*
  * DirectX type definitions
@@ -661,6 +664,13 @@ typedef enum {
 	CAM_Cinematic,
 	CAM_Heavy,
 } CAMERA_TYPE;
+
+typedef enum {
+	CFL_None,
+	CFL_FollowCenter,
+	CFL_NoChunky,
+	CFL_ChaseObject,
+} CAMERA_FLAG;
 
 typedef enum {
 	KM_Forward,
@@ -1556,31 +1566,6 @@ typedef struct Phd3dPos_t {
 	__int16 rotZ;
 } PHD_3DPOS;
 
-typedef struct CameraInfo_t {
-	GAME_VECTOR pos;
-	GAME_VECTOR target;
-	CAMERA_TYPE type;
-	int shift;
-	DWORD flags;
-	int fixedCamera;
-	int numberFrames;
-	int bounce;
-	int underwater;
-	int targetDistance;
-	int targetSquare;
-	__int16 targetAngle;
-	__int16 actualAngle;
-	__int16 targetElevation;
-	__int16 box;
-	__int16 number;
-	__int16 last;
-	__int16 timer;
-	__int16 speed;
-	LPVOID item;
-	LPVOID last_item;
-	OBJECT_VECTOR *fixed;
-} CAMERA_INFO;
-
 typedef struct ItemInfo_t {
 	int floor;
 	DWORD touchBits;
@@ -1615,6 +1600,31 @@ typedef struct ItemInfo_t {
 	UINT16 clear_body : 1;
 	UINT16 pad : 7;
 } ITEM_INFO;
+
+typedef struct CameraInfo_t {
+	GAME_VECTOR pos;
+	GAME_VECTOR target;
+	CAMERA_TYPE type;
+	int shift;
+	DWORD flags;
+	int fixedCamera;
+	int numberFrames;
+	int bounce;
+	int underwater;
+	int targetDistance;
+	int targetSquare;
+	__int16 targetAngle;
+	__int16 actualAngle;
+	__int16 targetElevation;
+	__int16 box;
+	__int16 number;
+	__int16 last;
+	__int16 timer;
+	__int16 speed;
+	ITEM_INFO *item;
+	ITEM_INFO *last_item;
+	OBJECT_VECTOR *fixed;
+} CAMERA_INFO;
 
 typedef struct CollSide_t {
 	int floor;