+++ specific/smain.cpp
@@ -55,9 +55,11 @@ extern double InvGUI_Scale;
 #endif // FEATURE_HUD_IMPROVED
 
 #ifdef FEATURE_BACKGROUND_IMPROVED
+static char PictureSuffix[32];
 extern DWORD InvBackgroundMode;
 extern DWORD PictureStretchLimit;
 extern bool LoadingScreensEnabled;
+extern bool RemasteredPixEnabled;
 #endif // FEATURE_BACKGROUND_IMPROVED
 
 #ifdef FEATURE_VIDEOFX_IMPROVED
@@ -76,6 +78,11 @@ extern char ScreenshotPath[MAX_PATH];
 #ifdef FEATURE_MOD_CONFIG
 extern bool BarefootSfxEnabled;
 #endif // FEATURE_MOD_CONFIG
+
+#ifdef FEATURE_AUDIO_IMPROVED
+extern double InventoryMusicMute;
+extern double UnderwaterMusicMute;
+#endif // FEATURE_AUDIO_IMPROVED
 
 #ifdef FEATURE_VIEW_IMPROVED
 extern bool PsxFovEnabled;
@@ -143,7 +150,16 @@ BOOL __cdecl GameMain() {
 	S_UpdateInput();
 	IsVidModeLock = true;
 #ifdef FEATURE_BACKGROUND_IMPROVED
-	if( !BGND2_LoadPicture("data\\legal.pcx", FALSE, FALSE) ) {
+	int res = -1;
+	if( *PictureSuffix ) {
+		char fname[256];
+		snprintf(fname, sizeof(fname), "data\\legal%s.pcx", PictureSuffix);
+		res = BGND2_LoadPicture(fname, FALSE, FALSE);
+	}
+	if( res ) {
+		res = BGND2_LoadPicture("data\\legal.pcx", FALSE, FALSE);
+	}
+	if( !res ) {
 		BGND2_ShowPicture(30, 90, 10, 2, TRUE);
 	}
 #else // FEATURE_BACKGROUND_IMPROVED
@@ -258,15 +274,24 @@ __int16 __cdecl TitleSequence() {
 		IsTitleLoaded = TRUE;
 	}
 
-#if defined(FEATURE_GOLD) && !defined(FEATURE_BACKGROUND_IMPROVED)
-	S_DisplayPicture(IsGold()?"data\\titleg.pcx":"data\\title.pcx", TRUE);
-#else // !defined(FEATURE_GOLD) && !defined(FEATURE_BACKGROUND_IMPROVED)
-	S_DisplayPicture("data\\title.pcx", TRUE);
-#endif // !defined(FEATURE_GOLD) && !defined(FEATURE_BACKGROUND_IMPROVED)
-
 #ifdef FEATURE_BACKGROUND_IMPROVED
-	// NOTE: title menu fade-in was absent in the original game
-	BGND2_ShowPicture(15, 0, 0, 0, FALSE);
+	int res = -1;
+	if( *PictureSuffix ) {
+		char fname[256];
+		snprintf(fname, sizeof(fname), "data\\title%s.pcx", PictureSuffix);
+		res = BGND2_LoadPicture(fname, TRUE, FALSE);
+	}
+	if( res ) {
+		res = BGND2_LoadPicture("data\\title.pcx", TRUE, FALSE);
+	}
+	if( !res ) {
+		// NOTE: title menu fade-in was absent in the original game
+		BGND2_ShowPicture(15, 0, 0, 0, FALSE);
+	}
+#elif defined(FEATURE_GOLD)
+	S_DisplayPicture(IsGold()?"data\\titleg.pcx":"data\\title.pcx", TRUE);
+#else // FEATURE_BACKGROUND_IMPROVED
+	S_DisplayPicture("data\\title.pcx", TRUE);
 #endif // FEATURE_BACKGROUND_IMPROVED
 
 	if( GF_GameFlow.titleTrack != 0 )
@@ -488,14 +513,16 @@ void __cdecl S_LoadSettings() {
 	GetRegistryBoolValue(REG_PSXBARPOS_ENABLE, &PsxBarPosEnabled, false);
 	GetRegistryFloatValue(REG_GAME_GUI_SCALE, &GameGUI_Scale, 1.0);
 	GetRegistryFloatValue(REG_INV_GUI_SCALE, &InvGUI_Scale, 1.0);
-	CLAMP(GameGUI_Scale, 1.0, 2.0);
-	CLAMP(InvGUI_Scale, 1.0, 2.0);
+	CLAMP(GameGUI_Scale, 0.5, 2.0);
+	CLAMP(InvGUI_Scale, 0.5, 2.0);
 #endif // FEATURE_HUD_IMPROVED
 
 #ifdef FEATURE_BACKGROUND_IMPROVED
 	GetRegistryDwordValue(REG_INVBGND_MODE, &InvBackgroundMode, 1);
 	GetRegistryDwordValue(REG_PICTURE_STRETCH, &PictureStretchLimit, 10);
+	GetRegistryBoolValue(REG_REMASTER_PIX_ENABLE, &RemasteredPixEnabled, true);
 	GetRegistryBoolValue(REG_LOADING_SCREENS, &LoadingScreensEnabled, false);
+	GetRegistryStringValue(REG_PICTURE_SUFFIX, PictureSuffix, sizeof(PictureSuffix), "");
 #endif // FEATURE_BACKGROUND_IMPROVED
 
 #ifdef FEATURE_VIDEOFX_IMPROVED
@@ -517,6 +544,13 @@ void __cdecl S_LoadSettings() {
 #ifdef FEATURE_ASSAULT_SAVE
 	GetRegistryBinaryValue(REG_GAME_ASSAULT, (LPBYTE)&Assault, sizeof(ASSAULT_STATS), NULL);
 #endif // FEATURE_ASSAULT_SAVE
+
+#ifdef FEATURE_AUDIO_IMPROVED
+	GetRegistryFloatValue(REG_INV_MUSIC_MUTE, &InventoryMusicMute, 1.0);
+	GetRegistryFloatValue(REG_UW_MUSIC_MUTE, &UnderwaterMusicMute, 1.0);
+	CLAMP(InventoryMusicMute, 0.0, 1.0);
+	CLAMP(UnderwaterMusicMute, 0.0, 1.0);
+#endif // FEATURE_AUDIO_IMPROVED
 
 #ifdef FEATURE_VIEW_IMPROVED
 	GetRegistryBoolValue(REG_PSXFOV_ENABLE, &PsxFovEnabled, false);